Refactor ConfigScreen to manage TeamViewer launch with alert confirmation and improve event handling

+++ campimetria/pages/config_screen.py
@@ -43,6 +43,7 @@ class ConfigScreen:
         self.cor_botao = game.cor_botao
         self.cor_botao_hover = game.cor_botao_hover
         self.cor_texto = game.cor_texto 
+        self.visualizando_alerta = False
  
 
 
@@ -58,22 +59,26 @@ class ConfigScreen:
                     if self.button_selected < 1:
                         self.button_selected += 1
                 elif event.key == pygame.K_e:  # Confirma a seleção
-                    if self.button_selected == 0:
-                        self.game.change_screen(testar_joystick.TestarJoystick(self.game))
-                    elif self.button_selected == 1:
-                        if subprocess.run(["pgrep", "teamviewer"], stdout=subprocess.PIPE):                            # subprocess.Popen([self.inicia_team])
-                            os.system("killall teamviewer")
-                        subprocess.Popen(["teamviewer", "--allowRoot"])                        
-                       
-                                      
-                      
-                
+                    if self.visualizando_alerta:
+                        if self.button_selected == 0:
+                            self.visualizando_alerta = False
+                        elif self.button_selected == 1:
+                            self.visualizando_alerta = False
+                            if subprocess.run(["pgrep", "teamviewer"], stdout=subprocess.PIPE):                            # subprocess.Popen([self.inicia_team])
+                                os.system("killall teamviewer")
+                            subprocess.Popen(["teamviewer", "--allowRoot"])                        
+                   
+                    else:       
+                        if self.button_selected == 0:
+                            self.game.change_screen(testar_joystick.TestarJoystick(self.game))
+                        elif self.button_selected == 1:
+                            self.visualizando_alerta = True                      
+                        
                 elif event.key == pygame.K_j:
                     self.game.change_screen(index.Index(self.game))
      
 
-    def update(self):
-       
+    def update(self):       
         pass
 
     def draw(self, surface):
@@ -89,7 +94,12 @@ class ConfigScreen:
         
         self.draw_button(surface, "TESTAR JOYSTICK", self.game.width // 2 - largura_botao//2 , 623 - espacamento,largura_botao, altura_botao, self.button_selected == 0)
         self.draw_buttonTeam(surface, "TEAMVIEWER",  self.game.width  // 2 - largura_botao//2, 623 ,largura_botao, altura_botao, self.button_selected == 1)
-        
+        if self.visualizando_alerta:
+            self.draw_alert(surface)
+            espacamento = 350
+            self.draw_button(surface, "CANCELAR",  self.game.width // 2 - largura_botao//2 - espacamento, 623,largura_botao, altura_botao, self.button_selected == 0)
+            self.draw_button(surface, "CONTINUAR",  self.game.width // 2 - largura_botao//2 + espacamento, 623 ,largura_botao, altura_botao, self.button_selected == 1)
+            
        
     def draw_button(self, surface, text,x, y, width, height, selected,visible = True):
         if visible:
@@ -109,3 +119,13 @@ class ConfigScreen:
             texto_render = self.font.render(text, True, cor_texto_selected if selected else cor_texto_unselected)
             text_rect = texto_render.get_rect(center=(x + width // 2, y + height // 2))       
             surface.blit(texto_render, text_rect)
+            
+    def draw_alert(self,surface):
+        largura_alerta = 1200
+        altura_alerta = 800
+        x = (self.game.width - largura_alerta) // 2
+        y = (self.game.height - altura_alerta) // 2 -100
+        pygame.draw.rect(surface, self.cor_fundo, (x, y, largura_alerta, altura_alerta), border_radius=10)
+        texto_renderizado = self.font.render("DESEJA ABRIR REALMENTE O TEAMVIEWER?", True, (255, 255, 255))
+        text_rect = texto_renderizado.get_rect(center=(x + largura_alerta // 2, y + altura_alerta // 2))
+        surface.blit(texto_renderizado, text_rect)
